Added multiple twitch user support

+++ cogs/update.py
@@ -1,6 +1,5 @@
-import discord
 import json
-import random
+import requests
 from discord.ext import commands
 
 class Update:
@@ -10,20 +9,35 @@ class Update:
     @commands.command()
     async def update(self, ctx):
 
-        with open('token.json') as talk1:
+        with open('twitch.json') as talk1:
             data1 = json.load(talk1)
             talk1.close()
 
-        channel = self.client.get_channel(int(data1['discord_twitch_channel']))
-        livemessage = await channel.get_message(int(data1['discord_message_id']))
-        await livemessage.delete()
+        streamers = data1['streamers']
+        twitch_client_id = data1['token']['Twitch_Client-ID']
+
+        for streamer in streamers:
+
+            stream = requests.get("https://api.twitch.tv/helix/streams?user_id={0}".format(streamer), headers=twitch_client_id).json()
+
+            if streamer == data1['streamers'][0]:
+                live_notification = data1['discord']['live_notification']['streamer1']
+                returndata_msgid = 'disc_msg_id_str1'
+                print("Streamer 1")
+            elif streamer == data1['streamers'][1]:
+                live_notification = data1['discord']['live_notification']['streamer2']
+                returndata_msgid = 'disc_msg_id_str2'
+                print("Streamer 1")
+
+            if live_notification == "alreadysend":
+                print("i am here now")
+                channel = self.client.get_channel(int(data1['discord']['discord_channel']))
+                livemessage = await channel.get_message(int(data1['discord']['live_notification']["{}".format(returndata_msgid)]))
+                print(livemessage)
+                await livemessage.delete()
 
         await ctx.send("Updating bot and restarting...")
         await self.client.close()
-
-
-    async def version(self, ctx):
-        await ctx.send("current version: super hyper pre alpha")
 
 def setup(client):
     client.add_cog(Update(client))
